Added `static_mem` implementation
This is a very important feature for static memory context.
